Move fetching of GitHub API client responsibility to deployment interface
