+++ config/github.go
@@ -1,0 +1,7 @@
+package config
+
+import (
+	"github.com/google/go-github/github"
+)
+
+var DefaultGithubClient *github.Client = nil
